cleaning up build output a bit
