+++ cetlvast/CMakeLists.txt
@@ -73,6 +73,7 @@ message(STATUS
     CMAKE_SOURCE_DIR:                ${CMAKE_SOURCE_DIR}\n\
     CMAKE_CURRENT_BINARY_DIR:        ${CMAKE_CURRENT_BINARY_DIR}\n\
     CMAKE_MESSAGE_LOG_LEVEL:         ${CMAKE_MESSAGE_LOG_LEVEL}\n\
+    CMAKE_INSTALL_PREFIX:            ${CMAKE_INSTALL_PREFIX}\n\
     CETLVAST_EXTERNAL_ROOT:          ${CETLVAST_EXTERNAL_ROOT}\n\
     FETCHCONTENT_FULLY_DISCONNECTED: ${FETCHCONTENT_FULLY_DISCONNECTED}\n\
 -----------------------------------------------------------\n\
@@ -158,8 +159,15 @@ include(${CETLVAST_FLAG_SET})
 # +---------------------------------------------------------------------------+
 cmake_path(APPEND CETL_ROOT "include" OUTPUT_VARIABLE CETL_INCLUDE)
 
-add_library(cetl INTERFACE)
-target_include_directories(cetl INTERFACE ${CETL_INCLUDE})
+include(ProjectLibrary)
+
+add_project_library(
+    NAME cetl
+    HEADER_PATH
+        ${CETL_INCLUDE}/
+    HEADER_INCLUDE_PATTERNS
+        *.hpp
+)
 
 find_package(clangformat)
 
@@ -209,7 +217,7 @@ if(DEFINED LOCAL_FORMAT_CHECK_TARGET_NAME)
             ${LOCAL_FORMAT_CHECK_TARGET_NAME}
     )
 else()
-    # no-op since there was not format target
+    # no-op since there was no format target
     add_custom_target(
         lint
     )
@@ -252,6 +260,9 @@ else()
     )
 endif()
 
+# "release" is a legacy target supported for older build systems that use this
+# as a default. It actually doesn't build the docs which is part of a CETL
+# release.
 add_custom_target(
     release
     DEPENDS
